requirements.txt
asgiref==3.7.2
bokeh==3.3.4
contourpy==1.2.0
cycler==0.12.1
Django==5.0.2
django-debug-toolbar==4.3.0
et-xmlfile==1.1.0
fonttools==4.49.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
nmrglue==0.10
numpy==1.26.4
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2024.1
PyYAML==6.0.1
scipy==1.12.0
six==1.16.0
sqlparse==0.4.4
tornado==6.4
typing_extensions==4.9.0
tzdata==2024.1
wcwidth==0.2.13
xyzservices==2023.10.1
psycopg
psycopg2
psycopg2-binary

